refactor: CheckDigitCalculatorVar2 を CheckDigitCalculator として上書き。
